requirements.txt
tensorflow
numpy
Pillow
faiss-cpu
pandas
tqdm
scikit-learn